Move `AgentSchema` to `stateful`
