+++ packages/engine/lib/stateful/src/agent/mod.rs
@@ -1,0 +1,3 @@
+mod schema;
+
+pub use self::schema::AgentSchema;
