+++ packages/engine/lib/stateful/src/lib.rs
@@ -1,5 +1,6 @@
 // TODO: DOC: Describing this crate containing stateful things like agents, context, etc. and the
 //            `field` interface
+pub mod agent;
 pub mod field;
 pub mod proxy;
