+++ packages/engine/src/datastore/table/state/create_remove/command.rs
@@ -1,11 +1,12 @@
 use std::{collections::HashSet, sync::Arc};
 
 use arrow::record_batch::RecordBatch;
+use stateful::agent::AgentSchema;
 
 use crate::{
     datastore::{
         error::{Error, Result},
-        schema::{state::AgentSchema, EngineComponent},
+        schema::EngineComponent,
         UUID_V4_LEN,
     },
     simulation::command::CreateRemoveCommands,
