+++ packages/engine/src/datastore/test_utils.rs
@@ -2,9 +2,12 @@ use std::sync::Arc;
 
 use rand::{prelude::StdRng, Rng, SeedableRng};
 use serde::{Deserialize, Serialize};
-use stateful::field::{
-    FieldScope, FieldSpec, FieldSpecMap, FieldType, FieldTypeVariant, RootFieldSpec,
-    RootFieldSpecCreator,
+use stateful::{
+    agent::AgentSchema,
+    field::{
+        FieldScope, FieldSpec, FieldSpecMap, FieldType, FieldTypeVariant, RootFieldSpec,
+        RootFieldSpecCreator,
+    },
 };
 use uuid::Uuid;
 
@@ -15,7 +18,7 @@ use crate::{
     },
     datastore::{
         error::Error,
-        schema::{last_state_index_key, state::AgentSchema, EngineComponent},
+        schema::{last_state_index_key, EngineComponent},
     },
     hash_types::state::{Agent, AgentStateField},
     proto::{ExperimentRunBase, InitialState, InitialStateName, ProjectBase},
