+++ packages/engine/src/simulation/command.rs
@@ -8,7 +8,7 @@ use std::{collections::HashSet, sync::Arc};
 
 use rayon::iter::{IndexedParallelIterator, IntoParallelIterator, ParallelIterator};
 use serde::{Deserialize, Serialize};
-use stateful::{field::RootFieldKey, proxy::PoolReadProxy};
+use stateful::{agent::AgentSchema, field::RootFieldKey, proxy::PoolReadProxy};
 use uuid::Uuid;
 
 use crate::{
@@ -18,7 +18,7 @@ use crate::{
             message::{CREATE_AGENT, REMOVE_AGENT, STOP_SIM},
         },
         batch::MessageBatch,
-        schema::{state::AgentSchema, EngineComponent},
+        schema::EngineComponent,
         table::{pool::message, references::MessageMap, state::create_remove::ProcessedCommands},
         UUID_V4_LEN,
     },
